# apps/desktop-companion/openapi.yaml
openapi: 3.1.0
info:
  title: FluxBoard Livestream Copilot API
  description: |
    REST API for the FluxBoard Livestream Copilot desktop companion service.

    FluxBoard provides real-time livestream capture, AI-powered content generation,
    OBS Studio integration, clip management, social media export, and video streaming
    for content creators.

    ## Authentication

    Most endpoints require authentication via one of two methods:

    1. **Bearer JWT** - Pass an access token in the `Authorization` header.
    2. **API Key** - Pass an API key in the `x-api-key` header.

    Public endpoints (health checks, OBS status, event reads) do not require authentication
    during development.

    ## WebSocket Events

    The server exposes a WebSocket on a configurable port (default 9001). Clients receive
    real-time `EventEnvelope` messages for session lifecycle, transcripts, clips, outputs,
    and trigger detections. Connect and listen for JSON messages.

    Event types: `SESSION_START`, `SESSION_END`, `TRANSCRIPT_SEGMENT`, `MOMENT_MARKER`,
    `CLIP_INTENT_START`, `CLIP_INTENT_END`, `ARTIFACT_CLIP_CREATED`, `ARTIFACT_FRAME_CREATED`,
    `OUTPUT_CREATED`, `OUTPUT_VALIDATED`, `AUTO_TRIGGER_DETECTED`, `CLIP_QUEUE_UPDATED`.

    ## Rate Limiting

    Rate-limited endpoints return standard `RateLimit-*` headers. When exceeded, the server
    responds with `429 Too Many Requests`.

    | Endpoint Group        | Limit              |
...

  version: 1.0.0
  contact:
...
  license:
...
servers:
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /api/v1/auth/login or /api/v1/auth/refresh
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Static API key for service-to-service communication
  schemas:
...
    Pagination:
      type: object
      properties:
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0
        total:
          type: integer
          example: 12

    # ── Auth ───────────────────────────────────────────────────────────────
...
    Event:
      type: object
      properties:
        id:
          type: string
        sessionId:
          type: string
        type:
          type: string
          enum:
            - SESSION_START
            - SESSION_END
            - TRANSCRIPT_SEGMENT
            - MOMENT_MARKER
            - CLIP_INTENT_START
            - CLIP_INTENT_END
            - ARTIFACT_CLIP_CREATED
            - ARTIFACT_FRAME_CREATED
            - OUTPUT_CREATED
            - OUTPUT_VALIDATED
            - AUTO_TRIGGER_DETECTED
            - CLIP_QUEUE_UPDATED
        payload:
          type: object
        ts:
          type: string
          description: Unix timestamp as string (BigInt)
        timestamp:
          type: number
          description: Unix timestamp as number
        traceId:
          type: string
          nullable: true
        spanId:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
...
  parameters:
...
    SessionIdParam:
      name: sessionId
      in: path
      required: true
      schema:
        type: string
      description: Session ID
...
    OffsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
        minimum: 0

# =============================================================================
# PATHS
# =============================================================================
paths:
...
  /api/events:
    get:
      tags: [Events]
      summary: List events with filtering
      operationId: listEvents
      security: []
      parameters:
        - name: sessionId
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: types
          in: query
          schema:
            type: string
          description: Comma-separated event types
        - name: afterTs
          in: query
          schema:
            type: string
          description: Filter events after this unix timestamp
        - name: beforeTs
          in: query
          schema:
            type: string
          description: Filter events before this unix timestamp
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
            maximum: 500
        - $ref: '#/components/parameters/OffsetParam'
        - name: orderDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Paginated event list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: object
                    properties:
                      events:
                        type: array
                        items:
                          $ref: '#/components/schemas/Event'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /api/events/{id}:
    get:
      tags: [Events]
      summary: Get event by ID
      operationId: getEvent
      security: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
        "404":
          description: Event not found
  /api/events/moments:
    get:
      tags: [Events]
      summary: List moment markers
      operationId: listMoments
      security: []
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        "200":
          description: Moment markers for session
    post:
      tags: [Events]
      summary: Create a moment marker
      operationId: createMoment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sessionId, label, timestamp]
              properties:
                sessionId:
                  type: string
                type:
                  type: string
                  enum: [hype, qa, sponsor, clip, highlight, marker]
                  default: marker
                label:
                  type: string
                  maxLength: 200
                description:
                  type: string
                  maxLength: 1000
                timestamp:
                  type: number
                  minimum: 0
                clipId:
                  type: string
            example:
              sessionId: cmlbcgemg0000i7b80ufgh3ko
              type: hype
              label: Big reaction
              timestamp: 342.5
      responses:
        "201":
          description: Moment created
        "404":
          description: Session not found
  /api/events/moments/{momentId}:
    delete:
      tags: [Events]
      summary: Delete a moment marker
      operationId: deleteMoment
      parameters:
        - name: momentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Moment deleted
        "400":
          description: Event is not a moment marker
        "404":
          description: Moment not found
  /api/sessions/{sessionId}/events:
    get:
      tags: [Sessions, Events]
      summary: List events for a session
      operationId: getSessionEvents
      security: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: type
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 100
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        "200":
          description: Session events
        "404":
          description: Session not found
  /api/sessions/{sessionId}/events/moments:
    get:
      tags: [Sessions, Events]
      summary: List moments for a session
      operationId: getSessionMoments
      security: []
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        "200":
          description: Session moments
    post:
      tags: [Sessions, Events]
      summary: Create moment for a session
      operationId: createSessionMoment
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label, timestamp]
              properties:
                type:
                  type: string
                  enum: [hype, qa, sponsor, clip, highlight, marker]
                label:
                  type: string
                description:
                  type: string
                timestamp:
                  type: number
                clipId:
                  type: string
      responses:
        "201":
          description: Moment created
  /api/sessions/{sessionId}/events/moments/{momentId}:
    delete:
      tags: [Sessions, Events]
      summary: Delete moment from a session
      operationId: deleteSessionMoment
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
        - name: momentId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Moment deleted
        "403":
          description: Moment does not belong to session

  # ── Outputs ──────────────────────────────────────────────────────────────
...